Move demo to docs and add start script
