+++ docs/.babelrc
