+++ docs/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "formidable-oss-landers-demo",
+  "name": "formidable-oss-landers-docs",
   "version": "0.1.0",
   "main": "static.config.js",
   "license": "MIT",
@@ -27,6 +27,7 @@
     "react-static-plugin-md-pages": "^0.1.7",
     "react-static-plugin-react-router": "^7.2.3",
     "react-static-plugin-sitemap": "^7.2.3",
+    "react-static-plugin-source-filesystem": "^7.4.2",
     "react-static-plugin-styled-components": "^7.2.2",
     "styled-components": "^5.1.0"
   },
