+++ docs/src/constants.js
@@ -2,5 +2,4 @@ export default {
   title: "demo",
   description: "",
   github: "https://www.github.com/FormidableLabs/formidable-oss-landers",
-  url: "https://formidable.com/open-source/demo",
 };
