+++ package.json
@@ -22,6 +22,7 @@
     "styled-components": "^5.1.1"
   },
   "scripts": {
+    "start": "cd docs && yarn start",
     "build": "rollup -c",
     "dev": "rollup -c -w"
   },
